star: mpu: Readd the old mpu3050 code
Since there are some problems with the gyro sensor, this is a try to bring back the old code.
In the 3.1 Kernel the mpu driver that is used is located in drivers/misc/inv_mpu. In the 2.6
Kernel the driver under drivers/misc/mpu3050 is used. This patch brings back the old functions and LG changes

Change-Id: I9dfb0840b3c468a14d3c3eafa90ce7d1d533f5a8