arm: tegra: Fix derp in the pm
See the comment in pm.h:

"The following two constants are for setting the CPU freq
floor when display is on. 204000Khz is for tablet and
102000KHz is for phones. The reason for different values
for tablet and phone is due to phones usually have smart
displays that requires less CPU activity for refreshing
the screen"

But in pm.c only the constant CPU_WAKE_FREQ_HIGH is used with
no check if it is a tablet or a phone. This commit uses
CPU_WAKE_FREQ_LOW for mach star since it is a phone.
That will also enable lower freq than 216MHz for future use

Change-Id: Ic163c6588259f42f31b6ddf88738bb59fdbc3c0c